refactor(eventstore): dedup cross-backend query logic (#518)
QueryBySession/QueryTurnStats/scanTurns 在 SQLiteStore 与 pgStore 间
复制粘贴约 150 行, 仅 query map / NULL 类型 / logger 差异.

提取到 query_shared.go:
- queryExecer 接口 (*sql.DB 与 *dbutil.DB 共用, 后者嵌入 *sql.DB)
- queryBySession: 游标分页共享实现 (65 行去重)
- turnScanDest: turn 行 22 列扫描定义共享 (防列序漂移)
- turnStatScanDest + collectTurnStats: turn 统计共享循环,
  success 列 (SQLite NullInt64 / PG NullBool) 通过回调抽象

零行为变更 (除统一 scanEvents 错误包装). build/vet/test 全通过.

Refs #518
